Allow fp16 without autocast

+++ ldm/modules/diffusionmodules/openaimodel.py
@@ -781,7 +781,7 @@ class UNetModel(nn.Module):
         for module in self.output_blocks:
             h = th.cat([h, hs.pop()], dim=1)
             h = module(h, emb, context)
-        h = h.type(x.dtype)
+        h = h.type(self.dtype)
         if self.predict_codebook_ids:
             return self.id_predictor(h)
         else:
